refactor: extract refresh dispatch and retry methods to PollingOperations
Move _dispatch_async_request_refresh, _schedule_short_retry, _handle_dr_event
(~78 lines) from main.py to polling.py.
