+++ custom_components/googlefindmy/coordinator/main.py
@@ -3622,84 +3622,6 @@ class GoogleFindMyCoordinator(
 
     # --- END: Add/Replace inside Coordinator class --------------------------------
 
-    # ---------------------------- Event loop helpers ------------------------
-    def _dispatch_async_request_refresh(
-        self, *, task_name: str, log_context: str
-    ) -> None:
-        """Invoke ``async_request_refresh`` safely regardless of its implementation."""
-
-        fn = getattr(self, "async_request_refresh", None)
-        if not callable(fn):
-            return
-
-        def _invoke() -> None:
-            try:
-                result = fn()
-                if inspect.isawaitable(result):
-                    self.hass.async_create_task(result, name=task_name)
-            except Exception as err:
-                _LOGGER.debug(
-                    "async_request_refresh dispatch failed (%s): %s", log_context, err
-                )
-
-        if self._is_on_hass_loop():
-            _invoke()
-        else:
-            self._run_on_hass_loop(_invoke)
-
-    def _schedule_short_retry(self, delay_s: float = 5.0) -> None:
-        """Schedule a short, coalesced refresh instead of shifting the poll baseline.
-
-        Rationale:
-        - When FCM/push is not ready, we *do not* advance `_last_poll_mono`.
-          Advancing the baseline hides readiness transitions and can put the
-          scheduler to "sleep". Instead, we request a short follow-up refresh.
-
-        Behavior:
-        - Coalesces multiple calls by cancelling a pending callback first.
-        - Always runs on the HA event loop.
-
-        Args:
-            delay_s: Delay in seconds before requesting a coordinator refresh.
-        """
-
-        def _do_schedule() -> None:
-            # Cancel a pending short retry (coalesce)
-            if self._short_retry_cancel is not None:
-                try:
-                    self._short_retry_cancel()
-                except Exception:  # defensive
-                    pass
-                finally:
-                    self._short_retry_cancel = None
-
-            def _cb(_now: datetime) -> None:
-                # Clear handle and request a refresh (non-blocking)
-                self._short_retry_cancel = None
-                self._dispatch_async_request_refresh(
-                    task_name=f"{DOMAIN}.short_retry_refresh",
-                    log_context="short retry",
-                )
-
-            self._short_retry_cancel = async_call_later(
-                self.hass, max(0.0, float(delay_s)), _cb
-            )
-
-        if self._is_on_hass_loop():
-            _do_schedule()
-        else:
-            self._run_on_hass_loop(_do_schedule)
-
-    # ---------------------------- Device Registry helpers -------------------
-    async def _handle_dr_event(self, _event: Event) -> None:
-        """Handle Device Registry changes by rebuilding poll targets (rare)."""
-        self._reindex_poll_targets_from_device_registry()
-        # After changes, request a refresh so the next tick uses the new target sets.
-        self._dispatch_async_request_refresh(
-            task_name=f"{DOMAIN}.dr_event_refresh",
-            log_context="device registry event",
-        )
-
     # ---------------------------- Cooldown helpers (server-aware) -----------
     def _compute_type_cooldown_seconds(self, report_hint: str | None) -> int:
         """Return a server-aware cooldown duration in seconds for a crowdsourced report type.
